+++ packages/doc-layout-component/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/settings-component
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 0.0.19
 
 ### Patch Changes